endpoint: Wait for proxy config completion before updating BPF maps
In case the Envoy proxy is run in a sidecar container, udpating the
BPF policy maps first may result in dropping the control traffic
to Istio Pilot required by Envoy to get configured and ACK the
policy configuration.

Wait for Envoy to get configured and ACK first, then update the BPF
maps.

Signed-off-by: Romain Lenglet <romain@covalent.io>